+++ src/ralph/schemas/edx/converters/xapi/base.py
@@ -1,0 +1,9 @@
+"""Base Xapi Converter"""
+
+from ralph.schemas.edx.converters.base_converter import BaseConverter
+
+
+class BaseXapiConverter(BaseConverter):
+    """Base Xapi Converter"""
+
+    _platform = ""